+++ android/src/main/kotlin/io/adrop/adrop_ads/banner/AdropBannerViewFactory.kt
@@ -8,9 +8,12 @@ import io.flutter.plugin.common.StandardMessageCodec
 import io.flutter.plugin.platform.PlatformView
 import io.flutter.plugin.platform.PlatformViewFactory
 
-class AdropBannerViewFactory(private val messenger: BinaryMessenger) : PlatformViewFactory(StandardMessageCodec.INSTANCE) {
+class AdropBannerViewFactory(
+    private val messenger: BinaryMessenger,
+    private val viewManager: AdropBannerManager
+) : PlatformViewFactory(StandardMessageCodec.INSTANCE) {
     override fun create(context: Context, viewId: Int, args: Any?): PlatformView {
         val callData = CallCreateBanner(args as? Map<String, Any?>)
-        return AdropBannerView(context, viewId, messenger, callData)
+        return AdropBannerView(context, viewId, messenger, callData, viewManager)
     }
 }