+++ example/integration_test/plugin_integration_test.dart
@@ -9,7 +9,7 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 
-import 'package:adrop_ads_flutter/adrop.dart';
+import 'package:adrop_ads_flutter/adrop_ads_flutter.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();