+++ ios/Classes/banner/AdropBannerView.swift
@@ -1,27 +1,23 @@
-//
-//  AdropBannerView.swift
-//  adrop_ads_flutter
-//
-//  Created by aaron on 2023/10/19.
-//
-
 import Foundation
 import Flutter
 import UIKit
 import AdropAds
 
 class AdropBannerView: NSObject, FlutterPlatformView, AdropBannerDelegate {
-    private var banner: AdropBanner
-    private var methodChannel: FlutterMethodChannel
-    
+    private let banner: AdropBanner
+    private let bannerManager: AdropBannerManager
+    private let methodChannel: FlutterMethodChannel
+
     init(
         frame: CGRect,
         viewIdentifier viewId: Int64,
         binaryMessenger messenger: FlutterBinaryMessenger,
-        call: CallCreateBanner
-    ) { 
-        banner = AdropBanner(unitId: call.unitId)
-        methodChannel = FlutterMethodChannel(name: AdropChannel.methodBannerChannelOf(id: viewId.description), binaryMessenger: messenger)
+        call: CallCreateBanner,
+        bannerManager: AdropBannerManager
+    ) {
+        self.bannerManager = bannerManager
+        self.methodChannel = FlutterMethodChannel(name: AdropChannel.methodBannerChannelOf(id: viewId.description), binaryMessenger: messenger)
+        self.banner = bannerManager.create(unitId: call.unitId)
 
         super.init()
         banner.delegate = self
@@ -31,7 +27,7 @@ class AdropBannerView: NSObject, FlutterPlatformView, AdropBannerDelegate {
     func view() -> UIView {
         return banner
     }
-    
+
     func onMethodCall(call: FlutterMethodCall, result: FlutterResult) {
         switch(call.method) {
         case AdropMethod.LOAD_BANNER:
@@ -41,16 +37,19 @@ class AdropBannerView: NSObject, FlutterPlatformView, AdropBannerDelegate {
             result(FlutterMethodNotImplemented)
         }
     }
-    
+
     func onAdReceived(_ banner: AdropBanner) {
         methodChannel.invokeMethod(AdropMethod.DID_RECEIVE_AD, arguments: nil)
+        bannerManager.onAdReceived(banner)
     }
-    
+
     func onAdClicked(_ banner: AdropBanner) {
         methodChannel.invokeMethod(AdropMethod.DID_CLICK_AD, arguments: nil)
+        bannerManager.onAdClicked(banner)
     }
-    
+
     func onAdFailedToReceive(_ banner: AdropBanner, _ error: AdropAds.AdropErrorCode) {
-        methodChannel.invokeMethod(AdropMethod.DID_FAILED_TO_RECEIVE, arguments: error.rawValue)
+        methodChannel.invokeMethod(AdropMethod.DID_FAILED_TO_RECEIVE, arguments: AdropErrorCodeToString(code: error))
+        bannerManager.onAdFailedToReceive(banner, error)
     }
 }