+++ ios/Classes/bridge/AdropChannel.swift
@@ -1,10 +1,3 @@
-//
-//  AdropChannel.swift
-//  adrop_ads_flutter
-//
-//  Created by aaron on 2023/10/19.
-//
-
 import Foundation
 
 struct AdropChannel {