+++ ios/Classes/model/CallCreateBanner.swift
@@ -1,10 +1,3 @@
-//
-//  CallCreateBanner.swift
-//  adrop_ads_flutter
-//
-//  Created by aaron on 2023/10/19.
-//
-
 import Foundation
 import AdropAds