+++ lib/src/adrop_method_channel.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/services.dart';
+
+import 'bridge/adrop_channel.dart';
+import 'bridge/adrop_method.dart';
+import 'adrop_platform_interface.dart';
+
+const _methodChannel = MethodChannel(AdropChannel.methodChannel);
+
+class AdropMethodChannel extends AdropPlatform {
+  @override
+  Future<void> initialize(bool production) async {
+    await _methodChannel.invokeMethod(AdropMethod.initialize, production);
+  }
+
+  @override
+  Future<void> loadBanner(String unitId) async {
+    await _methodChannel.invokeMethod(AdropMethod.loadBanner, unitId);
+  }
+}